cleanup, restore persistence
